% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{add_dim_red_seurat}
\alias{add_dim_red_seurat}
\title{Add dimensionality reduction to seurat.}
\usage{
add_dim_red_seurat(seurat_obj, dim_red_list, assay.used = "RNA",
  prefix = "")
}
\arguments{
\item{seurat_obj}{Data.}

\item{dim_red_list}{Assay in Seurat object.}

\item{assay.used}{.}

\item{prefix}{Boolean. To use variable features.}
}
\value{
Seurat obj.
}
\description{
Add dimensionality reduction to seurat.
}
