% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calc_clust_averages}
\alias{calc_clust_averages}
\title{Get cluster averages.}
\usage{
calc_clust_averages(metadata, data, group)
}
\arguments{
\item{metadata}{Metadata.}

\item{data}{Gene expression data.}

\item{group}{Column in metadata.}
}
\value{
.
}
\description{
Get cluster averages.
}
