% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{calculate_mito_pct}
\alias{calculate_mito_pct}
\title{Calculate mitochondrial percentage from Seurat object.}
\usage{
calculate_mito_pct(x)
}
\arguments{
\item{x}{A Seurat object.}
}
\value{
Seurat object.
}
\description{
Deprecated: renamed to \code{\link[=add_gene_class_percent]{add_gene_class_percent()}}, which also adds
\code{pct_ribo} and \code{pct_hb}.
}
\keyword{internal}
