% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_metadata_column}
\alias{check_metadata_column}
\title{Function to merge two metadata tables together.}
\usage{
check_metadata_column(metadata)
}
\arguments{
\item{metadata1}{A Seurat object or a tibble containing metadata with
either a column called "cell" with cell IDs or rownames with cell IDs.}

\item{metadata2}{A tibble containing metadata with
either a column called "cell" with cell IDs or rownames with cell IDs.}

\item{log_file}{A log filename.}

\item{proj_name}{Name of the project that will be the prefix of the file name.}

\item{label}{An optional label for the file.}

\item{write}{Boolean. To write out the metadata or not.}

\item{out_dir}{Directory in which to write metadata.}
}
\value{
A metadata file merged on cell identifiers.
}
\description{
Function to merge two metadata tables together.
}
