% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composition_barplot}
\alias{composition_barplot}
\title{Render bar plots summarizing celltype proportions and distribution in samples and groups}
\usage{
composition_barplot(
  scoot_object = NULL,
  sample_col = NULL,
  layer = "layer_1",
  return_plot_to_var = FALSE,
  facet_by = NULL
)
}
\arguments{
\item{scoot_object}{A scoot class object (typically after applying merge_scoot_objects onto a list of scootObjects)}

\item{layer}{Default "layer_1" if you have one cell type annotation layer in your scoot_object. Alternatively "layer_2" etc. if you have multiple layers of annotation depths.}

\item{return_plot_to_var}{Optionally, you can save the ggplots to a variable if you would like to further modify and adapt the plots on your own.}

\item{facet_by}{This allows you to pass a metadata column name present in your scoot_object$metadata to show your samples in facets with ggplot facet_grid, for example by "condition".}
}
\value{
Plotting function to show the cell type composition from scoot object across different samples.
}
\description{
Render bar plots summarizing celltype proportions and distribution in samples and groups
}
