% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composition_boxplot}
\alias{composition_boxplot}
\title{Render box plots summarizing celltype proportions and distribution in samples and groups}
\usage{
composition_boxplot(
  scoot_object = NULL,
  plot_var = "clr",
  layer = "layer_1",
  return_plot_to_var = FALSE,
  group_by = NULL,
  facet_by = NULL,
  pval_method = "wilcox.test",
  p_adjust_method = "BH",
  palette = "lancet",
  legend_position = "right"
)
}
\arguments{
\item{scoot_object}{A scoot class object (typically after applying merge_scoot_objects onto a list of scootObjects)}

\item{plot_var}{Column in the scoot_object$composition: either "freq" for cell type relative abundance in percent or "clr" (for centered log-ratio transformed). Default: "clr" as it is better suited for statistical analysis and is better able to also show low abundant cell types.}

\item{layer}{Default "layer_1" if you have one cell type annotation layer in your scoot_object. Alternatively "layer_2" etc. if you have multiple layers of annotation depths.}

\item{return_plot_to_var}{Optionally, you can save the ggplots to a variable if you would like to further modify and adapt the plots on your own.}

\item{group_by}{This allows you to pass a metadata column name present in your scoot_object$metadata to show your samples in groups, for example by "condition".}

\item{facet_by}{This allows you to pass a metadata column name present in your scoot_object$metadata to show your samples in facets with ggplot facet_grid, for example by "condition".}

\item{pval_method}{Specify method how to calculate the p-value. Wilcoxon test is recommended as compositional data might not fullfill the assumption of a gaussian distribution. For alternatives, see documentation of ggpubr::stat_pwc.}

\item{p_adjust_method}{Method for adjusting p-values (see ggpubr::stat_pwc for available methods)}

\item{palette}{Choose a palette of your liking. For available palettes, see ggsci package. Default: "lancet"}

\item{legend_position}{Where to put the legend. Possible options: "top", "right", "bottom", "left"}
}
\value{
Plotting function to show the cell type composition from scoot object across different samples.
}
\description{
Render box plots summarizing celltype proportions and distribution in samples and groups
}
