% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_color_vect}
\alias{create_color_vect}
\title{Function to create a color vector.}
\usage{
create_color_vect(seurat_obj, group = "orig.ident")
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{group}{Assay such as RNA.}
}
\value{
A vector of colors.
}
\description{
Function to create a color vector.
}
