% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{differential_expression_global}
\alias{differential_expression_global}
\title{Calculate differential expression for}
\usage{
differential_expression_global(
  data,
  metadata,
  metadata_column,
  log_fc_threshold = 0.5,
  min.pct = 0.1,
  test.use = "wilcox",
  out_path = ".",
  write = FALSE,
  log_file = NULL
)
}
\arguments{
\item{data}{Gene expression data.}

\item{metadata}{Metadata.}

\item{metadata_column}{Column in metadata.}

\item{log_fc_threshold}{Log fc threshold.}

\item{min.pct}{Minimum percentage of cells a gene must be expressed in to be tested.}

\item{test.use}{Test to use.}

\item{out_path}{output path.}

\item{write}{Boolean to write or not.}

\item{log_file}{log file.}
}
\value{
.
}
\description{
Calculate differential expression for
}
