% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{feature_corr_plot}
\alias{feature_corr_plot}
\title{Plot a heatmap showing feature correlations}
\usage{
feature_corr_plot(
  scores = NULL,
  min_samples = 10,
  ComplexHeatmap_args = list(col = rev(brewer.pal(n = 7, name = "RdYlBu")), rect_gp =
    grid::gpar(col = "grey", lwd = 1), column_names_max_height = unit(16, "cm"),
    row_names_max_width = unit(16, "cm"), heatmap_legend_param = list(title =
    "Correlation", legend_height = unit(4, "cm"), title_position = "lefttop-rot")),
  plot = TRUE
)
}
\arguments{
\item{scores}{Output from get_cluster_score}

\item{min_samples}{Minimum number of samples. If a feature space has less than min_samples, it will not be included in the heatmap plot}

\item{ComplexHeatmap_args}{List to provide arguments for ComplexHeatmap::Heatmap, see ?ComplexHeatmap::Heatmap for details.}
}
\value{
A heatmap showing feature correlations.
}
\description{
Feature correlation is determined by sample structure.
Imagine that sample (s) s1 is closest to s2, then s3, s4, ... based on cell type composition (feature space 1).
Based on pseudobulk (feature space 2) s1 is closest to s3, s2, s4, ... .
A signature (feature space 3) might order your samples as s1 closest to s10, s9, s8, ... .
So, cell type composition and pseudobulk show a very similar neighbourhood for sample s1, whereas the signature gives a very different neighbourhood structure.
Of course, all samples are taken into consideration to see which feature spaces structure your samples in a similar neighbourhood structure.
Technically, the distance matrix (dm) of each feature space is converted to ranks (to account for different scaling of feature space distances and for outliers).
For each dm separately, all rows are concatenated into one long dm vector (dm1 vector, dm2 vector, etc.).
For each dm vector the correlation to all other dm vectors are calculated.
The result is a correlation matrix, showing how similar each feature space structures the samples compared to all other feature spaces.
A high correlation = 1 between dm1 and dm2 means that they both structure samples in the way (each sample is close to the same neighbours in dm1 as in dm2)
}
