% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_score.R
\name{geneset_score}
\alias{geneset_score}
\title{Get geneset scores.}
\usage{
geneset_score(module_tbl, counts_raw, min_cpm = 0, limit_pct = 1)
}
\arguments{
\item{module_tbl}{geneset table.}

\item{counts_raw}{Raw counts}

\item{min_cpm}{.}

\item{limit_pct}{.}
}
\value{
.
}
\description{
Get geneset scores.
}
