% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_aggregated_profile}
\alias{get_aggregated_profile}
\title{Compute aggregated gene expression}
\usage{
get_aggregated_profile(
  object,
  group_by_aggregated = NULL,
  min_cells_aggregated = 10,
  assay = "RNA",
  useNA = FALSE,
  ...
)
}
\arguments{
\item{object}{A seurat object or a list of seurat objects}

\item{group_by_aggregated}{The Seurat object metadata column(s) containing celltype annotations}

\item{min_cells_aggregated}{Minimum number of cells per sample and cell type to aggregate data for pseudobulk or aggregated signatures. Aggregated data for cell types with less than that value will not be returned in aggregated data. Default value is 10.}

\item{assay}{Assay to retrieve information. By default "RNA".}

\item{useNA}{logical whether to return aggregated profile for NA (undefined) cell types, default is FALSE.}

\item{...}{Extra parameters for internal Seurat functions: AverageExpression, AggregateExpression, FindVariableFeatures}
}
\value{
Average and aggregated expression as a list of matrices for all genes and indicated gene lists filtering.
}
\description{
Function to compute aggregated expression (pseudobulk, i.e. sum counts per ident), and average expression by indicated cell type or grouping variable.
}
