% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_celltype_composition}
\alias{get_celltype_composition}
\title{Calculate cell type composition or frequencies}
\usage{
get_celltype_composition(
  object = NULL,
  group_by_composition = NULL,
  split_by = NULL,
  min_cells_composition = 10,
  useNA = FALSE,
  layer_links = c(scGate_multi = "functional.cluster"),
  clr_zero_impute_perc = 1,
  layer_1_link = "CellOntology_ID"
)
}
\arguments{
\item{object}{A seurat object}

\item{group_by_composition}{The Seurat object metadata column(s) containing celltype annotations (provide as character vector, containing the metadata column name(s))}

\item{split_by}{A Seurat object metadata column to split by (e.g. sample names)}

\item{min_cells_composition}{Set a minimum threshold for number of cells to calculate relative abundance (e.g. less than 10 cells -> no relative abundance will be calculated)}

\item{useNA}{Whether to include not annotated cells or not (labelled as "NA" in the group_by_composition). Can be defined separately for each group_by_composition (provide single boolean or vector of booleans)}

\item{layer_links}{Named vector indicating the relation of multiple layer classification, default is \code{c("scGate_multi", "functional.cluster")}. Name of the vector element ought to be layer_1 and element layer_2. This parameter is used to compute relative compositional data of layer_2 within layer_1 classes.}

\item{clr_zero_impute_perc}{To calculate the clr-transformed relative abundance ("clr_freq"), zero values are not allowed and need to be imputed (e.g. by adding a pseudo cell count). Instead of adding a pseudo cell count of flat +1, here a pseudo cell count of +1\% of the total cell count will be added to all cell types, to better take into consideration the relative abundance ratios (e.g. adding +1 cell to a total cell count of 10 cells would have a different, i.e. much larger effect, than adding +1 to 1000 cells).}

\item{layer_1_link}{Column of metadata linking layer_1 prediction (e.g. scGate ~ CellOntology_ID) in order to perform subsetting for second layer classification.}
}
\value{
Cell type compositions as a list of data.frames containing cell counts, relative abundance (freq) and clr-transformed freq (freq_clr), respectively.
}
\description{
Calculate cell type composition or frequencies
}
\examples{
devtools::install_github('satijalab/seurat-data')
library(SeuratData)
options(timeout = max(300, getOption("timeout")))
InstallData("panc8")
data("panc8")
panc8 = UpdateSeuratObject(object = panc8)
# Calculate overall composition
celltype_compositions.overall <- get_celltype_composition(object = panc8, group_by_composition = "celltype")

# Calculate sample-wise composition
celltype_compositions.sample_wise <- get_celltype_composition(object = panc8, group_by_composition = "celltype", split_by = "orig.ident")

}
