% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dr.R
\name{get_dr_point_size}
\alias{get_dr_point_size}
\title{Determine the point size for reduced dimensions scatter plots (smaller for
larger datasets).}
\usage{
get_dr_point_size(x)
}
\arguments{
\item{x}{Number of cells (points on the plot), a Seurat object, or a data
frame with one row per cell.}
}
\value{
Numeric point size.
}
\description{
Determine the point size for reduced dimensions scatter plots (smaller for
larger datasets).
}
