% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-data.R
\name{get_test_counts_matrix}
\alias{get_test_counts_matrix}
\title{Get an example counts matrix.}
\usage{
get_test_counts_matrix()
}
\value{
A matrix of raw counts.
}
\description{
Get a small matrix of raw counts from the PBMC dataset.
}
\examples{
pbmc_mat <- get_test_counts_matrix()

}
