% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_tsne_point_size}
\alias{get_tsne_point_size}
\title{Determine the point size for tSNE plots (smaller for larger datasets).}
\usage{
get_tsne_point_size(num_cells)
}
\arguments{
\item{num_cells}{Number of cells (points on the plot).}
}
\value{
Numeric point size.
}
\description{
Determine the point size for tSNE plots (smaller for larger datasets).
}
