% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{global_corr_plots}
\alias{global_corr_plots}
\title{Summarize scores and plot global correlation plot}
\usage{
global_corr_plots(scores, metadata)
}
\arguments{
\item{scores}{Output from get_cluster_scores}

\item{metadata}{The metadata from the merged scoot_object}
}
\value{
Two plots: 1. a correlation plot showing which clusterings and metadata columns are correlated. 2. correlation based on adjusted rand index.
}
\description{
Summarize scores and plot global correlation plot
}
