% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{import_matrix}
\alias{import_matrix}
\title{Read 10x data.}
\usage{
import_matrix(data_path, gene.column = 2)
}
\arguments{
\item{data_path}{path to 10x.}
}
\value{
list of matrices.
}
\description{
Read 10x data.
}
