% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_mtx}
\alias{import_mtx}
\title{Read in 10x Genomics Cell Ranger Matrix Market format data.}
\usage{
import_mtx(data_path, gene_column = 2, log_file = NULL)
}
\arguments{
\item{data_path}{Path to directory that holds the files output from 10x.}

\item{gene_column}{The column with the gene names.}

\item{log_file}{Filename for the log file.}
}
\value{
Named list of matrices. One matrix for each data type as specified
in the third column of the features.tsv file. As of Oct 3rd 2019, the two
options are \verb{Gene Expression} and \verb{Antibody Capture}
}
\description{
Read in 10x Genomics Cell Ranger Matrix Market format data.
}
