% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{load_sample_counts_matrix}
\alias{load_sample_counts_matrix}
\title{Read in Gene Expression and Antibody Capture data from a 10x Genomics Cell
Ranger sparse matrix or from a text file.}
\usage{
load_sample_counts_matrix(sample_name, path, log_file = NULL)
}
\arguments{
\item{sample_name}{A character that will be used as a prefix for all cell
names.}

\item{path}{Path to directory containing 10x matrix, or path to a text file.}

\item{log_file}{Filename for the log file.}
}
\value{
Named list of matrices. One matrix for each data type.
}
\description{
Deprecated: renamed to \code{\link[=read_counts_file]{read_counts_file()}}.
}
\keyword{internal}
