% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nas_per_sample}
\alias{nas_per_sample}
\title{Render plots summarizing celltype proportions and distribution in samples}
\usage{
nas_per_sample(
  obj_list = NULL,
  annot_col = c("scGate_multi"),
  return_plot = TRUE,
  bottom_mar = 10.2
)
}
\arguments{
\item{obj_list}{List of Seurat objects}

\item{annot_col}{Metadata column(s) containing the cell type annotations}

\item{bottom_mar}{Adjustable bottom margin for long sample names}
}
\value{
Get percentage of not annotated cells per sample and plot it.
}
\description{
Render plots summarizing celltype proportions and distribution in samples
}
