% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dr.R
\name{plot_dr_umap_clusters}
\alias{plot_dr_umap_clusters}
\title{Plot a UMAP colored by one or more cluster/grouping variables.}
\usage{
plot_dr_umap_clusters(
  x,
  cluster_var,
  color_scheme = get_color_scheme("clusters")
)
}
\arguments{
\item{x}{Seurat object with a "umap" reduction.}

\item{cluster_var}{Metadata column(s) to color by, as a single string
(comma-separated for more than one).}

\item{color_scheme}{(optional) Named vector of colors. Defaults to
\code{get_color_scheme("clusters")}.}
}
\value{
Nothing; called for its file output, one \verb{dr-umap-<cluster_var>.png}
per variable.
}
\description{
Plot a UMAP colored by one or more cluster/grouping variables.
}
