% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc.R
\name{plot_var_genes_upset}
\alias{plot_var_genes_upset}
\title{UpSet plot of the variable genes shared between batches.}
\usage{
plot_var_genes_upset(var_genes_list, nsets = 50, nintersects = 15)
}
\arguments{
\item{var_genes_list}{Named list of variable gene vectors, one element per
batch.}

\item{nsets}{Maximum number of sets to include.}

\item{nintersects}{Maximum number of intersections to show.}
}
\value{
An UpSet plot.
}
\description{
UpSet plot of the variable genes shared between batches.
}
