% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_counts_file}
\alias{read_counts_file}
\title{Read in Gene Expression and Antibody Capture data from a 10x Genomics Cell
Ranger sparse matrix or from a text file.}
\usage{
read_counts_file(path, sample_name, log_file = NULL)
}
\arguments{
\item{path}{Path to directory containing 10x matrix, or path to a text file.}

\item{sample_name}{A character that will be used as a prefix for all cell
names.}

\item{log_file}{Filename for the log file.}
}
\value{
Named list of matrices. One matrix for each data type. Currently the
only two data types are 'Gene Expression' and 'Antibody Capture'
}
\description{
Read in Gene Expression and Antibody Capture data from a 10x Genomics Cell
Ranger sparse matrix or from a text file.
}
