% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{read_remove.unmapped}
\alias{read_remove.unmapped}
\title{Reads in HTO data.}
\usage{
read_remove.unmapped(HTO_file)
}
\arguments{
\item{HTO_file}{Path to HTO file.}
}
\value{
hto matrix.
}
\description{
Reads in HTO data.
}
