% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_dimensionality_reduction}
\alias{run_dimensionality_reduction}
\title{Run dimensionality reduction, pca, tse, and umap}
\usage{
run_dimensionality_reduction(data, num_pcs, num_dim, num_neighbors,
  log_file = NULL, prefix = NULL)
}
\arguments{
\item{data}{A tibble with metadata.}

\item{num_pcs}{Minimum number of genes per cell.}

\item{num_dim}{Maximim number of genes per cell.}

\item{num_neighbors}{Maximum percentage of mitochondrial reads per cell.}

\item{log_file}{log file.}

\item{prefix}{suffix.}
}
\value{
list of dimensionality reduced/
}
\description{
Run dimensionality reduction, pca, tse, and umap
}
