% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scoot}
\alias{scoot}
\title{Get a scoot object summarizing cell type classification and aggregated profiles.}
\usage{
scoot(
  object,
  group_by = list(layer_1 = "scGate_multi", layer_2 = "functional.cluster"),
  split_by = NULL,
  min_cells_composition = 10,
  min_cells_aggregated = 10,
  name_additional_signatures = NULL,
  useNA = FALSE,
  clr_zero_impute_perc = 1,
  layer_links = c(scGate_multi = "functional.cluster"),
  assay = "RNA",
  layer_1_link = "CellOntology_ID",
  ncores = parallelly::availableCores() - 2,
  bparam = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{object}{A Seurat object or a list of Seurat objects}

\item{group_by}{List or vector with one or multiple Seurat object metadata columns with cell type predictions to group by (e.g. layer 1 cell type classification)}

\item{split_by}{A Seurat object metadata column to split by compositional data (e.g. sample names) \link{get_celltype_composition}}

\item{min_cells_composition}{Parameter for internal \link{get_celltype_composition}. Minimum number of cells annotated in a group_by parameter to render the cell type composition. If the number of cells annotated for a certain group_by parameter is less, compositional data will not be rendered. Default value is 10.}

\item{min_cells_aggregated}{Parameter for internal \link{get_aggregated_profile} and \link{get_aggregated_signature}. Minimum number of cells per sample and cell type to aggregate data for pseudobulk or aggregated signatures. Aggregated data for cell types with less than that value will not be returned in aggregated data. Default value is 10.}

\item{name_additional_signatures}{Names of additional signatures as found in object metadata to take into account.}

\item{useNA}{logical whether to return aggregated profile for NA (undefined) cell types, default is FALSE.}

\item{clr_zero_impute_perc}{Parameter for internal \link{get_celltype_composition}.}

\item{layer_links}{Parameter for internal \link{get_celltype_composition}}

\item{assay}{Parameter for internal \link{get_aggregated_profile}.}

\item{layer_1_link}{Column of metadata linking layer_1 prediction (e.g. scGate ~ CellOntology_ID) in order to perform subsetting for second layer classification.}
}
\value{
scoot object summarizing cell type classification and aggregated profiles.
}
\description{
Get a scoot object summarizing cell type classification and aggregated profiles.
}
