% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scooter-package.R
\docType{package}
\name{scooter-package}
\alias{scooter}
\alias{scooter-package}
\title{scooter: Streamlined scRNA-Seq Analysis Pipeline}
\description{
Streamlined scRNA-Seq analysis pipeline.
}
\author{
\strong{Maintainer}: Igor Dolgalev \email{igor.dolgalev@nyumc.org}

Authors:
\itemize{
  \item Igor Dolgalev \email{igor.dolgalev@nyumc.org}
  \item Anna Yeaton \email{ahyeaton@gmail.com}
}

}
\keyword{internal}
