% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{sctransform_data}
\alias{sctransform_data}
\title{SCT normalize data.}
\usage{
sctransform_data(umi, cell.attr, variable.features.n, log_file = NULL)
}
\arguments{
\item{log_file}{log file.}

\item{seurat_obj}{A seurat object.}
}
\value{
normalized data
}
\description{
SCT normalize data.
}
