% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seurat_to_matrix}
\alias{seurat_to_matrix}
\title{Function to extract data from Seurat object.}
\usage{
seurat_to_matrix(seurat_obj, assay = NULL, slot = NULL,
  reduction = NULL, metadata = TRUE)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{assay}{Assay such as RNA.}

\item{slot}{Slot such as counts.}

\item{reduction}{Character vector of reduction types.}

\item{metadata}{Boolean. To grab metadata or not}
}
\value{
A metadata file merged on cell identifiers.
}
\description{
Function to extract data from Seurat object.
}
