% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summarize_cluster_scores}
\alias{summarize_cluster_scores}
\title{Summarize scores and plot heat map}
\usage{
summarize_cluster_scores(
  data = NULL,
  topN = NULL,
  create_plots = TRUE,
  p_adjustment = TRUE,
  p_adjust_method = "fdr",
  p_value_cutoff = 0.05
)
}
\arguments{
\item{data}{Output from get_cluster_scores}

\item{topN}{Integer indicating number of topN highest scoring (most discriminating) features ranked for each score}

\item{create_plots}{Boolean indicating whether to create and show plots or not}

\item{p_adjustment}{Whether to adjust p-value columns or not}

\item{p_adjust_method}{Method for adjusting p-values (see stats::p.adjust for methods)}

\item{p_value_cutoff}{p-value (mean of all p-value columns) cutoff to filter out non-significant results}
}
\value{
Average silhouette widths per grouping variable. Optionally, a heatmap plot for visualization
}
\description{
Summarize scores and plot heat map
}
