% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{summarized_scores_barplot}
\alias{summarized_scores_barplot}
\title{Create a bar plot from summarized scores}
\usage{
summarized_scores_barplot(summarized_scores)
}
\arguments{
\item{summarized_scores}{Output from summarize_cluster_scores}
}
\value{
Summarize scores as bar plot
}
\description{
Create a bar plot from summarized scores
}
