% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_message}
\alias{write_message}
\title{Small function to write to message and to log file.}
\usage{
write_message(message_str, log_file = NULL)
}
\arguments{
\item{message_str}{A string to write as a message.}

\item{log_file}{A log filename.}
}
\value{
A message and writes the message to the specified log file.
}
\description{
Small function to write to message and to log file.
}
\examples{
write_message(message_str = "Finished Step 1", log_file = "log.file.txt")
}
