% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{zzz}
\alias{zzz}
\title{Add together two numbers}
\usage{
zzz(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
add(10, 1)
}
