% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{add_seurat_assay}
\alias{add_seurat_assay}
\title{Adds assay to Seurat object.}
\usage{
add_seurat_assay(seurat_obj, assay, counts_matrix, log_file = NULL)
}
\arguments{
\item{seurat_obj}{Existing seurat object.}

\item{assay}{Seurat assay to add the matrix to.}

\item{counts_matrix}{counts matrix.}

\item{log_file}{log file.}
}
\value{
Seurat object.
}
\description{
Adds assay to Seurat object.
}
