% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{add_seurat_assay}
\alias{add_seurat_assay}
\title{Add assay to Seurat object. If Seurat object doesn't exist, create one.}
\usage{
add_seurat_assay(seurat_obj, assay, counts_matrix, log_file = NULL)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{assay}{Assay slot.}

\item{counts_matrix}{counts matrix.}

\item{log_file}{log file.}
}
\value{
Seurat object.
}
\description{
Add assay to Seurat object. If Seurat object doesn't exist, create one.
}
