% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Run dimensionality reduction, pca, tse, and umap}
\usage{
calculate_clusters(pcs, num_dim, log_file, num_neighbors = 30,
  res = NULL)
}
\arguments{
\item{pcs}{Data.}

\item{num_dim}{Number of PCs to use for tsne and umap.}

\item{log_file}{log file.}

\item{num_neighbors}{Number of neighbors to use for umap.}

\item{res}{.}
}
\value{
.
}
\description{
Run dimensionality reduction, pca, tse, and umap
}
