% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{calculate_mito_pct}
\alias{calculate_mito_pct}
\title{Calculate mitochondrial percentage from Seurat object.}
\usage{
calculate_mito_pct(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object.}
}
\value{
Seurat object.
}
\description{
Calculate mitochondrial percentage from Seurat object.
}
