% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composition_boxplot}
\alias{composition_boxplot}
\title{Render box plots summarizing celltype proportions and distribution in samples and groups}
\usage{
composition_boxplot(
  scoot_object = NULL,
  plot_var = "clr",
  layer = "layer_1",
  return_plot_to_var = FALSE,
  group_by = NULL,
  facet_by = NULL,
  palette = "lancet",
  legend_position = "right"
)
}
\arguments{
\item{scoot_object}{A scoot class object (typically after applying merge_scoot_objects onto a list of scootObjects)}

\item{plot_var}{Column in the scoot_object$composition: either "freq" for cell type relative abundance in percent or "clr" (for centered log-ratio transformed). Default: "clr" as it is better suited for statistical analysis and is better able to also show low abundant cell types.}

\item{layer}{Default "layer_1" if you have one cell type annotation layer in your scoot_object. Alternatively "layer_2" etc. if you have multiple layers of annotation depths.}

\item{return_plot_to_var}{Optionally, you can save the ggplots to a variable if you would like to further modify and adapt the plots on your own.}

\item{group_by}{This allows you to pass a metadata column name present in your scoot_object$metadata to show your samples in groups, for example by "condition".}

\item{facet_by}{This allows you to pass a metadata column name present in your scoot_object$metadata to show your samples in facets with ggplot facet_grid, for example by "condition".}

\item{palette}{Choose a palette of your liking. For available palettes, see ggsci package. Default: "lancet"}

\item{legend_position}{Where to put the legend. Possible options: "top", "right", "bottom", "left"}

\item{p_adjust_method}{Method for adjusting p-values (see ggpubr::stat_pwc for available methods)}
}
\value{
Plotting function to show the cell type composition from scoot object across different samples.
}
\description{
Important note about significance testing: ggpubr::stat_compare_means is used for significance testing.
This means that p-values are not adjusted for multiple testing. Also, if there are more than two groups, per variable (cell type) it is tested whether there is a difference between the means (but not which one).
If any difference in means is found, this indicates to do subsequent pairwise comparisons to find which means are different.
}
