% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{create_seurat_obj}
\alias{create_seurat_obj}
\title{Create a new Seurat object from a matrix.}
\usage{
create_seurat_obj(
  counts_matrix,
  assay = "RNA",
  min_cells = 1,
  min_genes = 1,
  log_file = NULL,
  project = "proj"
)
}
\arguments{
\item{counts_matrix}{A matrix of raw counts.}

\item{assay}{Seurat assay to add the data to.}

\item{min_cells}{Include genes/features detected in at least this many cells.}

\item{min_genes}{Include cells where at least this many genes/features are detected.}

\item{log_file}{Filename for the logfile.}

\item{project}{Project name for Seurat object.}
}
\value{
Seurat object.
}
\description{
Create a new Seurat object from a matrix.
}
