% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{create_seurat_obj}
\alias{create_seurat_obj}
\title{Create a new Seurat object from a matrix.}
\usage{
create_seurat_obj(counts_matrix, assay = "RNA", log_file = NULL)
}
\arguments{
\item{counts_matrix}{A matrix of raw counts.}

\item{assay}{Seurat assay to add the data to.}

\item{log_file}{log file.}
}
\value{
Seurat object.
}
\description{
Create a new Seurat object from a matrix.
}
