% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{create_seurat_obj}
\alias{create_seurat_obj}
\title{Create a new Seurat object from a matrix.}
\usage{
create_seurat_obj(counts_matrix, min_cells = 10, min_genes = 100,
  out_dir = ".", color_scheme = NULL)
}
\arguments{
\item{counts_matrix}{A matrix of raw counts.}

\item{min_cells}{X.}

\item{min_genes}{X.}

\item{out_dir}{Directory where all the output files will be saved.}

\item{color_scheme}{A character vector of colors for plots (per sample/library).}
}
\value{
Seurat object.
}
\description{
Create a new Seurat object from a matrix.
}
\examples{
create_seurat_obj(counts_matrix = counts_matrix)

}
