% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_data}
\alias{filter_data}
\title{Filter cells based on the number of genes and mitochondrial reads.}
\usage{
filter_data(
  data,
  log_file = NULL,
  min_genes = NULL,
  max_genes = NULL,
  max_mt = 10
)
}
\arguments{
\item{data}{A tibble with metadata.}

\item{log_file}{Log file.}

\item{min_genes}{Minimum number of genes per cell.}

\item{max_genes}{Maximim number of genes per cell.}

\item{max_mt}{Maximum percentage of mitochondrial reads per cell.}
}
\value{
Filtered data
}
\description{
Filter out cells based on minimum and maximum number of genes and maximum percentage mitochondrial reads. If cutoffs are not provided, the min_genes will be the 0.02 quantile, and the max genes will be 0.98 quantile and the mitochondrial percentage will be 10%.
}
