% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_cluster_score}
\alias{get_cluster_score}
\title{Get metrics of cell types pseudobulk clustering}
\usage{
get_cluster_score(
  scoot_object = NULL,
  cluster_by = NULL,
  min_samples = 5,
  cluster_by_drop_na = TRUE,
  batching = NULL,
  scores = c("silhouette_isolated", "silhouette", "modularity"),
  dist_method = "euclidean",
  modularity_k = 3,
  max_nc = 5,
  pam_nclusters = "auto",
  hdbscan_min_pts = 3,
  NbClust_method = "ward.D2",
  ntests = 100,
  seed = 22,
  pval_combine_method = "weighted_zmethod",
  pca_comps_labs_invisible = c("quali"),
  pca_pb_labs_invisible = c("quali"),
  pca_sig_labs_invisible = c("quali"),
  ndim = 10,
  nvar_genes = 500,
  black_list = NULL,
  pca_n_hvg = 20,
  ncores = round(parallelly::availableCores()/2),
  bparam = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{scoot_object}{A scoot class object obtained with \link{scoot_object} or pseudobulk raw count matrix (\code{scootObject@aggregated_profile$pseudobulk$layer_1})}

\item{cluster_by}{Vector indicating the variable for clustering, default is celltype (for the annotation) and sample}

\item{min_samples}{Minimum number of samples to calculate scores (do not go lower than 5)}

\item{cluster_by_drop_na}{Whether to keep (FALSE) or drop (TRUE) NAs present in cluster_by column.}

\item{batching}{Vector indicating the variable for batching to allow calculating scores per batch, to account for batch effect}

\item{scores}{scores to compute clustering of samples based on cell type prediction.}

\item{dist_method}{Method to compute distance between celltypes, default is euclidean.}

\item{modularity_k}{Number of k-nearest neighbors to use to build graph for the calculation of the modularity score}

\item{pam_nclusters}{Number of clusters for PAM clustering. Default: 2 clusters. "auto" uses factoextra::fviz_nbclust(FUNcluster = cluster::pam, method = "silhouette") to suggest a number of clusters.}

\item{hdbscan_min_pts}{integer; Minimum size of clusters. See dbscan::hdbscan for details.}

\item{NbClust_method}{Hierarchical clustering method for fastNbClust (improved version of NbClust::NbClust).}

\item{ntests}{Number of shuffling events to calculate p-value for scores}

\item{seed}{Set seed for random shuffling events to calculate p-value for scores}

\item{pca_comps_labs_invisible}{Parameter to pass to fviz_pca defining which labels to plot for composition PCA, see documentation of fviz_pca for details.}

\item{pca_pb_labs_invisible}{Parameter to pass to fviz_pca defining which labels to plot for pseudobulk PCA, see documentation of fviz_pca for details.}

\item{pca_sig_labs_invisible}{Parameter to pass to fviz_pca defining which labels to plot for signatures PCA, see documentation of fviz_pca for details.}

\item{ndim}{Number of dimensions to be use for PCA clustering metrics. Default is 10.}

\item{nvar_genes}{Number of variable genes to assess samples. Default is 500.}

\item{black_list}{List of genes to discard from clustering, if "default" object "default_black_list" object is used. Alternative black listed genes can be provided as a vector or list.}

\item{pca_n_hvg}{Number of most highly variable genes to plot on the PCA (for pseudobulks).}

\item{ncores}{The number of cores to use, by default, all available cores - 2.}

\item{bparam}{A \code{BiocParallel::bpparam()} object that tells how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{progressbar}{Whether to show a progressbar or not}

\item{pval.combine}{Method for combining p-values if calculated using batching. Default is "zmethod" weighted (Stouffer's) Z-method, weighting by sample size per batch. Alternatively, Fisher's method can be used with "fisher".}
}
\value{
Metrics of cell types pseudobulk clustering
}
\description{
Get metrics of cell types pseudobulk clustering
}
