% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dr.R
\name{get_color_scheme}
\alias{get_color_scheme}
\title{Determine the color scheme.}
\usage{
get_color_scheme(type = "clusters")
}
\arguments{
\item{type}{Type of scheme ("samples" or "clusters").}
}
\value{
A vector of colors.
}
\description{
Determine the color scheme. Can be specified for samples or clusters to avoid confusion.
}
