% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-dr.R
\name{get_dr_point_size}
\alias{get_dr_point_size}
\title{Determine the point size for reduced dimensions scatter plots (smaller for larger datasets).}
\usage{
get_dr_point_size(num_cells)
}
\arguments{
\item{num_cells}{Number of cells (points on the plot) or a Seurat object.}
}
\value{
Numeric point size.
}
\description{
Determine the point size for reduced dimensions scatter plots (smaller for larger datasets).
}
