% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{load_sample_counts_matrix}
\alias{load_sample_counts_matrix}
\title{Reads in data from 10x, or from a delimited file.}
\usage{
load_sample_counts_matrix(sample_name, data_path_10x = NULL,
  AC_file = NULL, delim = NULL, log_file = NULL)
}
\arguments{
\item{sample_name}{A character for the sample name to be a prefix for every cell.}

\item{data_path_10x}{A path to the 10x data /<data_path>/outs.}

\item{AC_file}{A path to the antibody capture file.}

\item{delim}{A delimiter for the AC_file.}

\item{log_file}{A log filename.}
}
\value{
Named list of matrices. One matrix for each data type.
}
\description{
Reads in data from 10x, or from a delimited file.
}
