% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_sample_counts_matrix}
\alias{load_sample_counts_matrix}
\title{Reads in data from 10x, or from a delimited file.}
\usage{
load_sample_counts_matrix(sample_name, path, log_file = NULL)
}
\arguments{
\item{sample_name}{A character for the sample name to be a prefix for every cell.}

\item{path}{A path to the 10x data /<data_path>/outs.}

\item{log_file}{A log filename.}
}
\value{
Named list of matrices. One matrix for each data type.
}
\description{
Reads in data from 10x, or from a delimited file.
}
