% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{log_normalize_data}
\alias{log_normalize_data}
\title{Log normalize data.}
\usage{
log_normalize_data(data, log_file = NULL)
}
\arguments{
\item{log_file}{log file.}

\item{seurat_obj}{A seurat object.}

\item{assay}{assay.}
}
\value{
normalized data
}
\description{
Log normalize data.
}
