% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Function to merge two metadata tables together.}
\usage{
merge_metadata(metadata_one, metadata_two, log_file = NULL,
  write = TRUE, proj_name = "", label = "", out_dir = ".")
}
\arguments{
\item{metadata_one}{A Seurat object or a tibble containing metadata with
either a column called "cell" with cell IDs or rownames with cell IDs.}

\item{metadata_two}{A tibble containing metadata with
either a column called "cell" with cell IDs or rownames with cell IDs.}

\item{log_file}{A log filename.}

\item{write}{Boolean. To write out the metadata or not.}

\item{proj_name}{Name of the project that will be the prefix of the file name.}

\item{label}{An optional label for the file.}

\item{out_dir}{Directory in which to write metadata.}
}
\value{
A metadata file merged on cell identifiers.
}
\description{
Function to merge two metadata tables together.
}
