% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_metadata}
\alias{merge_metadata}
\title{Function to merge two metadata tables together.}
\usage{
merge_metadata(metadata1, metadata2, log_file = NULL)
}
\arguments{
\item{metadata1}{A Seurat object or a tibble containing metadata with
either a column called "cell" with cell IDs or rownames with cell IDs.}

\item{metadata2}{A tibble containing metadata with
either a column called "cell" with cell IDs or rownames with cell IDs.}

\item{log_file}{A log filename.}
}
\value{
A metadata file merged on cell identifiers.
}
\description{
Function to merge two metadata tables together.
}
