% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-normalize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data}
\usage{
normalize_data(data, normalize_method, nfeatures = 2000,
  metadata = NULL, assay = NULL, log_file = NULL)
}
\arguments{
\item{data}{Input data.}

\item{normalize_method}{.}

\item{nfeatures}{.}

\item{metadata}{.}

\item{assay}{.}

\item{log_file}{log file.}
}
\value{
Normalized data
}
\description{
Normalize data
}
