% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the distribution of specified features/variables.}
\usage{
plot_distribution(data, features, grouping, color_scheme = NULL)
}
\arguments{
\item{data}{Seurat object or metadata.}

\item{features}{Vector of features to plot (such as "nGene", "nUMI", "percent.mito").}

\item{grouping}{X.}

\item{color_scheme}{(optional) Vector of colors.}
}
\value{
A vector of colors.
}
\description{
Plot the distribution of specified features/variables.
}
