% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_dimensionality_reduction}
\alias{run_dimensionality_reduction}
\title{Run dimensionality reduction, pca, tse, and umap}
\usage{
run_dimensionality_reduction(data, assay = "RNA", var_features = TRUE,
  num_pcs, num_dim, num_neighbors, log_file = NULL, prefix = NULL)
}
\arguments{
\item{data}{Data.}

\item{num_pcs}{Number of PCs to calculate.}

\item{num_dim}{Number of PCs to use for tsne and umap.}

\item{num_neighbors}{Number of neighbors to use for umap.}

\item{log_file}{log file.}

\item{prefix}{prefix for dimensionality reduction labels.}
}
\value{
list of dimensionality reductions.
}
\description{
Run dimensionality reduction, pca, tse, and umap
}
