% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_dr}
\alias{run_dr}
\title{Run dimensionality reduction, pca, tse, and umap}
\usage{
run_dr(data, dr_method, ...)
}
\arguments{
\item{data}{A tibble with metadata.}

\item{dr_method}{.}

\item{assay}{.}

\item{features}{.}

\item{graph}{.}

\item{num_dim}{.}
}
\value{
list of dimensionality reduced/
}
\description{
Run dimensionality reduction, pca, tse, and umap
}
