% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_dr}
\alias{run_dr}
\title{Run dimensionality reduction, pca, tse, and umap}
\usage{
run_dr(data, dr_method, prefix, assay = NULL, var_features = FALSE,
  features = NULL, graph = NULL, num_dim_use = NULL,
  reduction = NULL, num_neighbors = NULL, num_pcs = NULL, ...)
}
\arguments{
\item{data}{Data to use for dimensionality reduction.}

\item{dr_method}{Dimensionality reduction method.}

\item{prefix}{.}

\item{assay}{.}

\item{var_features}{.}

\item{features}{.}

\item{graph}{.}

\item{num_dim}{.}

\item{min_dist}{.}
}
\value{
list of dimensionality reduced/
}
\description{
Run dimensionality reduction, pca, tse, and umap
}
\seealso{
run_pca(), run_tsne(), run_umap()
}
