% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA}
\usage{
run_pca(data, num_pcs, prefix = "PC_")
}
\arguments{
\item{data}{A tibble with metadata.}

\item{num_pcs}{Maximim number of genes per cell.}

\item{prefix}{suffix.}
}
\value{
named list of feature loadings, cell embeddings, sdev, output from pca
}
\description{
Run PCA
}
