% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality_reduction.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA}
\usage{
run_pca(data, num_dim, prefix = "PC_")
}
\arguments{
\item{data}{A tibble with metadata.}

\item{num_dim}{Maximim number of genes per cell.}

\item{prefix}{suffix.}
}
\value{
pca.
}
\description{
Run PCA
}
