% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_tsne}
\alias{run_tsne}
\title{Run TSNE}
\usage{
run_tsne(data, seed.use = 1, dim.embed = 2, prefix = "tSNE_")
}
\arguments{
\item{data}{A tibble with metadata.}

\item{seed.use}{.}

\item{dim.embed}{.}

\item{prefix}{suffix.}
}
\value{
pca.
}
\description{
Run TSNE
}
