% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_tsne}
\alias{run_tsne}
\title{Run TSNE}
\usage{
run_tsne(data, seed.use = 1, dim.embed = 2, prefix = "tSNE_")
}
\arguments{
\item{data}{Data to run tsne on.}

\item{seed.use}{seed to use.}

\item{dim.embed}{Number of tsne embeddings to return.}

\item{prefix}{suffix.}
}
\value{
tsne.
}
\description{
Run TSNE
}
