% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_umap}
\alias{run_umap}
\title{Run UMAP}
\usage{
run_umap(data, num_neighbors, min_dist = 0.3, prefix = "UMAP_")
}
\arguments{
\item{data}{Data to run UMAP on.}

\item{num_neighbors}{Number of neighbors.}

\item{min_dist}{Distance metric.}

\item{prefix}{Prefix.}
}
\value{
umap
}
\description{
Run UMAP
}
