% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimentionality-reduction.R
\name{run_umap}
\alias{run_umap}
\title{Run UMAP}
\usage{
run_umap(data, num_neighbors, prefix = "UMAP_")
}
\arguments{
\item{data}{Data to run UMAP on.}

\item{num_neighbors}{num neighbors.}

\item{prefix}{suffix.}
}
\value{
umap
}
\description{
Run UMAP
}
