% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scoot}
\alias{scoot}
\title{Get a scoot object summarizing cell type classification and aggregated profiles.}
\usage{
scoot(
  object,
  ann_layer_cols = list(layer_1 = "scGate_multi", layer_2 = "functional.cluster"),
  min_cells = 10,
  name_additional_signatures = c("IFN_UCell", "HeatShock_UCell", "cellCycle.G1S_UCell",
    "cellCycle.G2M_UCell"),
  useNA = FALSE,
  clr_zero_impute_perc = 1,
  layer_links = c(scGate_multi = "functional.cluster"),
  assay = "RNA",
  layer_1_link = "CellOntology_ID",
  ncores = parallelly::availableCores() - 2,
  bparam = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{object}{A Seurat object or a list of Seurat objects.}

\item{ann_layer_cols}{List or vector with one or multiple Seurat object metadata columns with cell type annotations (e.g. layer 1 broad cell type annotation (CD4, CD8, ...) and layer 2 with high-resolution subtypes (CD4.treg, CD4.tfh, CD8.naive, CD8.exhausted, ...))}

\item{min_cells}{Minimum number of cells for any annotated cell type in layer_1.}

\item{name_additional_signatures}{Names of additional signatures as found in object metadata to take into account.}

\item{useNA}{logical whether to return aggregated profile for NA (undefined) cell types.}

\item{clr_zero_impute_perc}{Parameter for internal \link{get_celltype_composition}.}

\item{layer_links}{Parameter for internal \link{get_celltype_composition}}

\item{assay}{Parameter for internal \link{get_aggregated_profile}.}

\item{layer_1_link}{Column of metadata linking layer_1 prediction (e.g. scGate ~ CellOntology_ID) in order to perform subsetting for second layer classification.}
}
\value{
scoot object summarizing cell type classification and aggregated profiles.
}
\description{
Get a scoot object summarizing cell type classification and aggregated profiles.
}
