% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{sctransform_data}
\alias{sctransform_data}
\title{SCT normalize data.}
\usage{
sctransform_data(seurat_obj, log_file = NULL)
}
\arguments{
\item{seurat_obj}{A seurat object.}

\item{log_file}{log file.}
}
\value{
normalized data
}
\description{
SCT normalize data.
}
