% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{sctransform_data}
\alias{sctransform_data}
\title{SCT normalize data.}
\usage{
sctransform_data(counts, metadata, nfeatures, log_file = NULL)
}
\arguments{
\item{counts}{Raw counts.}

\item{metadata}{Metadata for each cell.}

\item{nfeatures}{Number of variable features to output.}

\item{log_file}{A log file.}
}
\value{
A named list of the vst output, the final scaled data, and the top variable genes.
}
\description{
SCT normalize data.
}
