% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_identity}
\alias{set_identity}
\title{Set identity of the Seurat object.}
\usage{
set_identity(seurat_obj, identity_column)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{identity_column}{The name of the identity column to pull from object metadata.}
}
\value{
A Seurat object with an updated identity set.
}
\description{
Wrapper for SeuratObject::Idents() with extra safety checks.
}
